Fix ansible-lint warnings during ansible-galaxy import
This commit fixes the ansible-lint warnings that was affecting
the quality score of aoscx role in ansible-galaxy.

The commit adds platform name to the role info.
It also fixes the role tags to only contain lowercase alphabets.
It also fixes the sample playbook to show the correct rolename
after installation through ansible-galaxy commands.